% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{aside}
\alias{aside}
\title{Add term to margin and index but not text}
\usage{
aside(margin = margin, index = margin)
}
\arguments{
\item{margin}{Text to show in margin}

\item{index}{Text to show in index}
}
\description{
Add term to margin and index but not text
}
