% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached.R
\name{assign_cached}
\alias{assign_cached}
\title{Assign a cached version of a function in the nanespace}
\usage{
assign_cached(fn_name, package, cachedir)
}
\arguments{
\item{fn_name}{(\code{character(1)})/cr
The name of the function to overload with a cached version.}

\item{package}{(\code{character(1)})/cr
The package in which the  function should be overloaded, e.g. \code{"mlr3oml"}.}

\item{cachedir}{(\code{character(1)})/cr
The cache directory relative to \code{.../mlr3book/book}.}
}
\description{
Create a cached version of a function and overload the original function in the package namespace.
Currently this is used to ensure that the CI for the OpenML chapter does not fail when the server is down.
}
