% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_mlr3book}
\alias{clean_mlr3book}
\title{Clean the mlr3 book}
\usage{
clean_mlr3book(clean = FALSE)
}
\arguments{
\item{clean}{Passed down to \code{\link[bookdown:clean_book]{bookdown::clean_book()}}.}
}
\description{
Calls \code{\link[bookdown:clean_book]{bookdown::clean_book()}}.
}
