% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{define}
\alias{define}
\title{Add term to index and margin}
\usage{
define(main, margin = NULL, index = NULL, code = FALSE)
}
\arguments{
\item{main}{Text to show in book}

\item{margin}{Text to show in margin}

\item{index}{Text to show in index}

\item{code}{If TRUE tells function to wrap in ``}
}
\description{
Add term to index and margin
}
