% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_sales_prediction.R
\docType{data}
\name{house_sales_prediction}
\alias{house_sales_prediction}
\title{Predicting House Prices}
\format{A data.table with 19 house features plus the price and the id columns, 21613 observations.
'}
\source{
https://www.kaggle.com/harlfoxem/housesalesprediction
License: CC0: Public Domain
}
\usage{
house_sales_prediction
}
\description{
This dataset contains house sale prices for King County, which includes Seattle.
It includes homes sold between May 2014 and May 2015.
It's a great dataset for evaluating simple regression models.
}
\examples{
data("house_sales_prediction", package = "mlr3book")
}
\keyword{datasets}
