% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_book_stable.R
\name{install_mlr3verse_stable}
\alias{install_mlr3verse_stable}
\title{Install mlr3verse of last successful mlr3book build}
\usage{
install_mlr3verse_stable(dry = FALSE)
}
\arguments{
\item{dry}{(\code{logical(1)})\cr
When \code{dry = TRUE}, only the Github hashes of the packages are returned but no installation is done.}
}
\description{
Install mlr3verse of last successful mlr3book build
}
\details{
Installs the mlr3verse package versions from the last successful build of the mlr3book.
This is useful when the Travis CI build of the mlr3book is broken and you still want to compile it for writing.
}
