% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{link}
\alias{link}
\title{Create markdown and print-friendly link}
\usage{
link(url, text = NULL)
}
\arguments{
\item{url}{URL to link to}

\item{text}{Text to display in main text}
}
\description{
Creates markdown link and footnote with full link
}
