% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mlr3verse_stable.R
\name{list_mlr3verse_stable}
\alias{list_mlr3verse_stable}
\title{Lists the mlr3verse pkg states of last successful mlr3book build}
\usage{
list_mlr3verse_stable()
}
\description{
Lists the mlr3verse pkg states of last successful mlr3book build
}
