% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_chunks_mlr3book}
\alias{name_chunks_mlr3book}
\title{Name all chunks}
\usage{
name_chunks_mlr3book(file = NULL)
}
\arguments{
\item{file}{\code{character(1)}\cr
Optional path to a single qmd file. If \code{NULL}, all qmd files in the \verb{book/} directory are processed.}
}
\description{
Names all chunks of all qmd files using the pattern \verb{[file-name]-[number]}.
}
