% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_chunks_mlrbook}
\alias{name_chunks_mlrbook}
\title{Name all chunks}
\usage{
name_chunks_mlrbook()
}
\description{
Names all chunks of all Rmd files using the pattern \code{[file-name]-[number]}.
}
