% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_mlr3book}
\alias{render_mlr3book}
\title{Render the mlr3 book}
\usage{
render_mlr3book(format = "html", ...)
}
\arguments{
\item{format}{(\code{character()}).
Vector of output formats. Possible values: \code{"html"}, \code{"pdf"}.}

\item{...}{Passed down to \code{\link[bookdown:render_book]{bookdown::render_book()}}.}
}
\description{
Calls \code{\link[bookdown:render_book]{bookdown::render_book()}}.
}
