% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{cran_pkg}
\alias{cran_pkg}
\title{Hyperlink to CRAN Package}
\usage{
cran_pkg(pkg)
}
\arguments{
\item{pkg}{Name of the CRAN package.}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Creates a markdown link to a CRAN package.
}
