% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{gh_pkg}
\alias{gh_pkg}
\title{Hyperlink to GitHub Repository}
\usage{
gh_pkg(pkg)
}
\arguments{
\item{pkg}{Name of the repository specified as "{repo}/{name}".}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Creates a markdown link to GitHub repository.
}
