% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{mlr_pkg}
\alias{mlr_pkg}
\title{Hyperlink to mlr3 Package}
\usage{
mlr_pkg(pkg)
}
\arguments{
\item{pkg}{Name of the mlr3 package.}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Creates a markdown link to a mlr3 package with a "mlr-org.com" subdomain.
}
