% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_chunks_mlr3book}
\alias{name_chunks_mlr3book}
\title{Name all chunks}
\usage{
name_chunks_mlr3book()
}
\description{
Names all chunks of all qmd files using the pattern \verb{[file-name]-[number]}.
}
