% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{ref}
\alias{ref}
\title{Hyperlink to Function Reference}
\usage{
ref(topic, text = NULL, format = "markdown", aside = FALSE)
}
\arguments{
\item{topic}{Name of the topic to link against.}

\item{text}{Text to use for the link. Defaults to the topic name.}

\item{format}{Either markdown or HTML.}

\item{aside}{If \code{TRUE} calls \code{aside(topic)}}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Creates a markdown link to a function reference.
}
