% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{ref}
\alias{ref}
\title{Hyperlink to Function Reference}
\usage{
ref(topic, text = topic)
}
\arguments{
\item{topic}{Name of the topic to link against.}

\item{text}{Text to use for the link. Defaults to the topic name.}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Creates a markdown link to a function reference.
}
