% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{ref}
\alias{ref}
\title{Hyperlink to Function Reference}
\usage{
ref(topic, index = FALSE, aside = FALSE)
}
\arguments{
\item{topic}{Name of the topic to link against.}

\item{index}{If \code{TRUE} calls \code{index}}

\item{aside}{Passed to \code{index}}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Creates a markdown link to a function reference.
}
