% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{ref_pkg}
\alias{ref_pkg}
\title{Hyperlink to Package}
\usage{
ref_pkg(pkg, format = "markdown")
}
\arguments{
\item{pkg}{Name of the package.}

\item{format}{Either markdown or HTML.}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Links either to respective mlr3 website or to CRAN page.
}
