% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{ref_pkg}
\alias{ref_pkg}
\title{Hyperlink to Package}
\usage{
ref_pkg(pkg, runiverse = TRUE, index = FALSE, aside = FALSE)
}
\arguments{
\item{pkg}{Name of the package.}

\item{runiverse}{If \code{TRUE} (default) then creates R-universe link instead of GH}

\item{index}{If \code{TRUE} calls \code{index}}

\item{aside}{Passed to \code{index}}
}
\value{
(\code{character(1)}) markdown link.
}
\description{
Links either to respective mlr3 website or to CRAN page.
}
