% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_mlr3book}
\alias{render_mlr3book}
\title{Render the mlr3 book}
\usage{
render_mlr3book(...)
}
\arguments{
\item{...}{Passed down to \code{\link[bookdown:render_book]{bookdown::render_book()}}.}
}
\description{
Calls \code{\link[bookdown:render_book]{bookdown::render_book()}}.
}
